requirements.txt
python-dotenv
openai
ipykernel
ipywidgets
matplotlib
azure-ai-vision
diffusers
transformers
scipy
ftfy
accelerate
azure-ai-ml
azure-identity
mlflow
azureml-mlflow
tenacity
tqdm
faiss